backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pdf2image==1.16.3
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pytesseract==0.3.10
langdetect==1.0.9
requests==2.31.0
httpx==0.25.1
pandas==2.1.4
python-dateutil==2.8.2
redis==5.0.1
celery==5.3.4
scikit-learn==1.3.2
sentence-transformers==2.2.2